username and action audit implemented
(cherry picked from commit 87e9df05104b4bf7f7bcb742f233bbddcc527947)
