+++ src/Skoruba.IdentityServer4.Admin.EntityFramework.DbContexts/AuditEntry.cs
@@ -1,0 +1,63 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using Microsoft.EntityFrameworkCore.ChangeTracking;
+using Newtonsoft.Json;
+
+namespace Skoruba.IdentityServer4.Admin.EntityFramework.DbContexts
+{
+    /// <summary>
+    /// https://www.meziantou.net/2017/08/14/entity-framework-core-history-audit-table
+    /// https://dejanstojanovic.net/aspnet/2018/november/tracking-data-changes-with-entity-framework-core/
+    /// </summary>
+
+    public class AuditEntry
+    {
+        public AuditEntry(EntityEntry entry)
+        {
+            Entry = entry;
+        }
+
+        public EntityEntry Entry { get; }
+        public string TableName { get; set; }
+        public Dictionary<string, object> KeyValues { get; } = new Dictionary<string, object>();
+        public Dictionary<string, object> OldValues { get; } = new Dictionary<string, object>();
+        public Dictionary<string, object> NewValues { get; } = new Dictionary<string, object>();
+        public List<PropertyEntry> TemporaryProperties { get; } = new List<PropertyEntry>();
+
+        public bool HasTemporaryProperties => TemporaryProperties.Any();
+
+        public ClientConfigAudit ToAudit()
+        {
+            var audit = new ClientConfigAudit();
+            audit.TableName = TableName;
+            audit.DateTime = DateTime.UtcNow;
+            audit.KeyValues = JsonConvert.SerializeObject(KeyValues);
+            audit.OldValues = OldValues.Count == 0 ? null : JsonConvert.SerializeObject(OldValues);
+            audit.NewValues = NewValues.Count == 0 ? null : JsonConvert.SerializeObject(NewValues);
+            return audit;
+        }
+    }
+
+    public static class CollectionExtensions
+    {
+        public static IDictionary<TKey, TValue> NullIfEmpty<TKey, TValue>(this IDictionary<TKey, TValue> dictionary)
+        {
+            if (dictionary == null || !dictionary.Any())
+            {
+                return null;
+            }
+            return dictionary;
+        }
+
+        public static IEnumerable<T> ForEach<T>(this IEnumerable<T> source, Action<T> action)
+        {
+            foreach (T element in source)
+            {
+                action(element);
+            }
+            return source;
+        }
+    }
+
+}
