+++ src/Skoruba.IdentityServer4.Admin.EntityFramework.DbContexts/IdentityServerConfigurationDbContext.cs
@@ -17,9 +17,22 @@ namespace Skoruba.IdentityServer4.Admin.EntityFramework.DbContexts
 {
     public class IdentityServerConfigurationDbContext : ConfigurationDbContext<IdentityServerConfigurationDbContext>, IAdminConfigurationDbContext
     {
+
+        private readonly ILoggerFactory _loggerFactory;
+        private readonly IHttpContextAccessor _httpContextAccessor;
+
+
+        public IdentityServerConfigurationDbContext(DbContextOptions<IdentityServerConfigurationDbContext> options, ConfigurationStoreOptions storeOptions, ILoggerFactory loggerFactory, IHttpContextAccessor httpContextAccessor)
+            : base(options, storeOptions)
+        {
+            this._loggerFactory = loggerFactory;
+            this._httpContextAccessor = httpContextAccessor;
+        }
+
         public IdentityServerConfigurationDbContext(DbContextOptions<IdentityServerConfigurationDbContext> options, ConfigurationStoreOptions storeOptions)
             : base(options, storeOptions)
         {
+
         }
 
         public DbSet<ApiResourceProperty> ApiResourceProperties { get; set; }
@@ -67,188 +80,16 @@ namespace Skoruba.IdentityServer4.Admin.EntityFramework.DbContexts
                 audit.Property(x => x.OldValues).HasMaxLength(2048);
                 audit.Property(x => x.TableName).HasMaxLength(128).IsRequired();
                 audit.Property(x => x.DateTime).IsRequired();
+                audit.Property(x => x.Action).HasMaxLength(64);
+                audit.Property(x => x.Username).HasMaxLength(128);
             });
 
             base.OnModelCreating(modelBuilder);
         }
 
-
-        public override async Task<int> SaveChangesAsync(bool acceptAllChangesOnSuccess, CancellationToken cancellationToken = default(CancellationToken))
-        {
-            var auditEntries = OnBeforeSaveChanges();
-            var result = await base.SaveChangesAsync(acceptAllChangesOnSuccess, cancellationToken);
-            await OnAfterSaveChanges(auditEntries);
-            return result;
-        }
-
-        private List<AuditEntry> OnBeforeSaveChanges()
-        {
-            ChangeTracker.DetectChanges();
-            var auditEntries = new List<AuditEntry>();
-            foreach (var entry in ChangeTracker.Entries())
-            {
-                if (entry.Entity is ClientConfigAudit || entry.State == EntityState.Detached || entry.State == EntityState.Unchanged)
-                    continue;
-
-                var auditEntry = new AuditEntry(entry);
-                auditEntry.TableName = entry.Metadata.Relational().TableName;
-                auditEntries.Add(auditEntry);
-
-                foreach (var property in entry.Properties)
-                {
-                    if (property.IsTemporary)
-                    {
-                        // value will be generated by the database, get the value after saving
-                        auditEntry.TemporaryProperties.Add(property);
-                        continue;
-                    }
-
-                    string propertyName = property.Metadata.Name;
-                    if (property.Metadata.IsPrimaryKey())
-                    {
-                        auditEntry.KeyValues[propertyName] = property.CurrentValue;
-                        continue;
-                    }
-
-                    switch (entry.State)
-                    {
-                        case EntityState.Added:
-                            auditEntry.NewValues[propertyName] = property.CurrentValue;
-                            break;
-
-                        case EntityState.Deleted:
-                            auditEntry.OldValues[propertyName] = property.OriginalValue;
-                            break;
-
-                        case EntityState.Modified:
-
-                            var origin = property.OriginalValue != null ? property.OriginalValue.ToString() : "";
-                            var current = property.CurrentValue != null ? property.CurrentValue.ToString() : "";
-                            //if (!origin.Equals(current))
-                            //{
-                            auditEntry.OldValues[propertyName] = property.OriginalValue;
-                            auditEntry.NewValues[propertyName] = property.CurrentValue;
-                            //}
-
-                            break;
-                    }
-                }
-            }
-
-            // Save audit entities that have all the modifications
-            foreach (var auditEntry in auditEntries.Where(_ => !_.HasTemporaryProperties))
-            {
-                Audits.Add(auditEntry.ToAudit());
-            }
-
-            // keep a list of entries where the value of some properties are unknown at this step
-            return auditEntries.Where(_ => _.HasTemporaryProperties).ToList();
-        }
-
-        private Task OnAfterSaveChanges(List<AuditEntry> auditEntries)
-        {
-            if (auditEntries == null || auditEntries.Count == 0)
-                return Task.CompletedTask;
-
-            foreach (var auditEntry in auditEntries)
-            {
-                // Get the final value of the temporary properties
-                foreach (var prop in auditEntry.TemporaryProperties)
-                {
-                    if (prop.Metadata.IsPrimaryKey())
-                    {
-                        auditEntry.KeyValues[prop.Metadata.Name] = prop.CurrentValue;
-                    }
-                    else
-                    {
-                        auditEntry.NewValues[prop.Metadata.Name] = prop.CurrentValue;
-                    }
-                }
-
-                // Save the Audit entry
-                Audits.Add(auditEntry.ToAudit());
-            }
-
-            return SaveChangesAsync();
-        }
-
-    }
-
-    /// <summary>
-    /// https://www.meziantou.net/2017/08/14/entity-framework-core-history-audit-table
-    /// </summary>
-    public class AuditEntry
-    {
-        public AuditEntry(EntityEntry entry)
-        {
-            Entry = entry;
-        }
-
-        public EntityEntry Entry { get; }
-        public string TableName { get; set; }
-        public Dictionary<string, object> KeyValues { get; } = new Dictionary<string, object>();
-        public Dictionary<string, object> OldValues { get; } = new Dictionary<string, object>();
-        public Dictionary<string, object> NewValues { get; } = new Dictionary<string, object>();
-        public List<PropertyEntry> TemporaryProperties { get; } = new List<PropertyEntry>();
-
-        public bool HasTemporaryProperties => TemporaryProperties.Any();
-
-        public ClientConfigAudit ToAudit()
-        {
-            var audit = new ClientConfigAudit();
-            audit.TableName = TableName;
-            audit.DateTime = DateTime.UtcNow;
-            audit.KeyValues = JsonConvert.SerializeObject(KeyValues);
-            audit.OldValues = OldValues.Count == 0 ? null : JsonConvert.SerializeObject(OldValues);
-            audit.NewValues = NewValues.Count == 0 ? null : JsonConvert.SerializeObject(NewValues);
-            return audit;
-        }
-    }
-
-    public class AuditBase
-    {
-        public int Id { get; set; }
-        public string TableName { get; set; }
-        public DateTime DateTime { get; set; }
-        public string KeyValues { get; set; }
-        public string OldValues { get; set; }
-        public string NewValues { get; set; }
-        public String Username { get; set; }
-        public String Action { get; set; }
-    }
-
-    public class ClientConfigAudit : AuditBase
-    {
-
-    }
-
-   // https://dejanstojanovic.net/aspnet/2018/november/tracking-data-changes-with-entity-framework-core/
-    public class CatalogDbContext : DbContext
-    {
-        private readonly ILoggerFactory loggerFactory;
-        private readonly IHttpContextAccessor httpContextAccessor;
-
-        public CatalogDbContext(DbContextOptions<CatalogDbContext> options)
-         : base(options)
-        {
-        }
-
-        public CatalogDbContext(DbContextOptions<CatalogDbContext> options, ILoggerFactory loggerFactory, IHttpContextAccessor httpContextAccessor)
-          : base(options)
-        {
-            this.loggerFactory = loggerFactory;
-            this.httpContextAccessor = httpContextAccessor;
-        }
-
         protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
         {
-            optionsBuilder.UseLoggerFactory(loggerFactory);
-        }
-
-        protected override void OnModelCreating(ModelBuilder builder)
-        {
-            base.OnModelCreating(builder);
-            //builder.ApplyConfiguration(new ProductConfiguration());
+            optionsBuilder.UseLoggerFactory(_loggerFactory);
         }
 
         public override async Task<int> SaveChangesAsync(bool acceptAllChangesOnSuccess, CancellationToken cancellationToken = default(CancellationToken))
@@ -258,7 +99,6 @@ namespace Skoruba.IdentityServer4.Admin.EntityFramework.DbContexts
             await AuditTemporaryProperties(temoraryAuditEntities);
             return result;
         }
-
 
         async Task<IEnumerable<Tuple<EntityEntry, ClientConfigAudit>>> AuditNonTemporaryProperties()
         {
@@ -271,7 +111,7 @@ namespace Skoruba.IdentityServer4.Admin.EntityFramework.DbContexts
                     TableName = e.Metadata.Relational().TableName,
                     Action = Enum.GetName(typeof(EntityState), e.State),
                     DateTime = DateTime.Now.ToUniversalTime(),
-                    Username = this.httpContextAccessor?.HttpContext?.User?.Identity?.Name,
+                    Username = this._httpContextAccessor?.HttpContext?.User?.Identity?.Name,
                     KeyValues = JsonConvert.SerializeObject(e.Properties.Where(p => p.Metadata.IsPrimaryKey()).ToDictionary(p => p.Metadata.Name, p => p.CurrentValue).NullIfEmpty()),
                     NewValues = JsonConvert.SerializeObject(e.Properties.Where(p => e.State == EntityState.Added || e.State == EntityState.Modified).ToDictionary(p => p.Metadata.Name, p => p.CurrentValue).NullIfEmpty()),
                     OldValues = JsonConvert.SerializeObject(e.Properties.Where(p => e.State == EntityState.Deleted || e.State == EntityState.Modified).ToDictionary(p => p.Metadata.Name, p => p.OriginalValue).NullIfEmpty())
@@ -280,17 +120,17 @@ namespace Skoruba.IdentityServer4.Admin.EntityFramework.DbContexts
 
             //Return list of pairs of EntityEntry and ToolAudit  
             return entitiesToTrack.Where(e => e.Properties.Any(p => p.IsTemporary))
-                 .Select(e => new Tuple<EntityEntry, ClientConfigAudit>(
-                     e,
-                 new ClientConfigAudit()
-                 {
-                     TableName = e.Metadata.Relational().TableName,
-                     Action = Enum.GetName(typeof(EntityState), e.State),
-                     DateTime = DateTime.Now.ToUniversalTime(),
-                     Username = this.httpContextAccessor?.HttpContext?.User?.Identity?.Name,
-                     NewValues = JsonConvert.SerializeObject(e.Properties.Where(p => !p.Metadata.IsPrimaryKey()).ToDictionary(p => p.Metadata.Name, p => p.CurrentValue).NullIfEmpty())
-                 }
-                 )).ToList();
+                .Select(e => new Tuple<EntityEntry, ClientConfigAudit>(
+                    e,
+                    new ClientConfigAudit()
+                    {
+                        TableName = e.Metadata.Relational().TableName,
+                        Action = Enum.GetName(typeof(EntityState), e.State),
+                        DateTime = DateTime.Now.ToUniversalTime(),
+                        Username = this._httpContextAccessor?.HttpContext?.User?.Identity?.Name,
+                        NewValues = JsonConvert.SerializeObject(e.Properties.Where(p => !p.Metadata.IsPrimaryKey()).ToDictionary(p => p.Metadata.Name, p => p.CurrentValue).NullIfEmpty())
+                    }
+                )).ToList();
         }
 
         async Task AuditTemporaryProperties(IEnumerable<Tuple<EntityEntry, ClientConfigAudit>> temporatyEntities)
@@ -298,36 +138,14 @@ namespace Skoruba.IdentityServer4.Admin.EntityFramework.DbContexts
             if (temporatyEntities != null && temporatyEntities.Any())
             {
                 await Audits.AddRangeAsync(
-                temporatyEntities.ForEach(t => t.Item2.KeyValues = JsonConvert.SerializeObject(t.Item1.Properties.Where(p => p.Metadata.IsPrimaryKey()).ToDictionary(p => p.Metadata.Name, p => p.CurrentValue).NullIfEmpty()))
-                    .Select(t => t.Item2)
+                    temporatyEntities.ForEach(t => t.Item2.KeyValues = JsonConvert.SerializeObject(t.Item1.Properties.Where(p => p.Metadata.IsPrimaryKey()).ToDictionary(p => p.Metadata.Name, p => p.CurrentValue).NullIfEmpty()))
+                        .Select(t => t.Item2)
                 );
                 await SaveChangesAsync();
             }
             await Task.CompletedTask;
         }
-
-        //public DbSet<Product> Products { get; set; }
-        public DbSet<ClientConfigAudit> Audits { get; set; }
     }
 
-    public static class Extensions
-    {
-        public static IDictionary<TKey, TValue> NullIfEmpty<TKey, TValue>(this IDictionary<TKey, TValue> dictionary)
-        {
-            if (dictionary == null || !dictionary.Any())
-            {
-                return null;
-            }
-            return dictionary;
-        }
 
-        public static IEnumerable<T> ForEach<T>(this IEnumerable<T> source, Action<T> action)
-        {
-            foreach (T element in source)
-            {
-                action(element);
-            }
-            return source;
-        }
-    }
 }
